Split desktop projection and journey test modules
