+++ crates/defra-agent-desktop/src/app/tests/live/operator_config/mod.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+mod apply;
+mod baseline;
+mod config;
+mod replication;
+
+pub(crate) use apply::apply_live_switch_config_in_operator;
+pub(crate) use baseline::assert_live_operator_switching_baseline;
+pub(crate) use config::{prepare_live_switch_config, LiveSwitchConfig};
+pub(crate) use replication::wait_for_live_switch_config_replication;
